rename otf.compile -> otf.compilation for consistency
